Make element getter/setter and add primitive method
The element value is now a getter setter that sets a private property on the
instance. This changes it so the element value may be set on the constructor,
but if it is not set, it defaults back to generic a "element" value.

It also adds a method on each instance that provides the primitive value from
which the class was either subclassed or registered.

Lastly, this converts functions in tests to arrow functions.