Add a name getter on meta.
Looks like this was partially removed in a previous commit, but the setter
was still there. What do you think of adding it back in? If not, we can just
remove the setter as well.

This is still used by HTTP headers in the resource namespace, but I can add
the getter/setter in Fury instead if no other elements will use a name in
their meta.