refactor: Don't create meta/attributes when checking if empty
The `meta` and `attributes` properties of an element are computed, in
the case that meta and attributes are not set when access. An empty
`ObjectElement` is returned.

When we check if meta or attributes is empty, if we access via the
getter there may be an unnecessary instantiation of an ObjectElement
which slows down serialisation when meta or attributes is not set.